PosList performance optimization Add new operating mode for matching a whole chunk
In validate, if a complete chunk matches, the PosList will not be
materialized with all ChunkOffsets from 0 to chunk->size(). Rather, the
PosList will store a shared_ptr to the chunk and only materialize if
necessary. This allows to speed up validate if a complete chunk is
valid. More optimizations that will reduce the materialization when
accessing the PosList to come.